requirements.txt
# Spare Part Demand Forecasting
# Core Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1.4

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Streamlit Dashboard
streamlit>=1.30.0

# Azure ML
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
azureml-mlflow>=1.55.0
mlflow>=2.10.0

# API & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.28.0
